+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Change Log
 ### Features
 * Notifica responsável pela abertura que o socorro foi atendido (PL-2541)
 * Cria método de upload de fotos da abertura para o S3 (PL-2518)
+* Cria relatório de dados gerais de socorros em rotas (PL-2523)
 
 <a name="v1.0.18"></a>
 ## Version [v1.0.18](https://github.com/luizfp/PrologWebService/compare/v1.0.17...v1.0.18) (release-date) [unreleased]